requirements.txt
line-bot-sdk==3.16.0
flask==3.0.0